requirements.txt
PyYAML==5.1.1
click==7.0
setuptools==41.0.1
paramiko==2.4.2
configobj==5.0.6
PyYAML==5.1.1